# src/main/resources/api/api.yml
swagger: "2.0"
info:
  description: "Here will be more information usefull for front-end developers"
  version: "0.0.1 ALPHA"
  title: "RPG API"
  contact:
    email: "jurasz.do@gmail.com"
host: "localhost"
basePath: "/"
...
schemes:
- "http"
paths:
  /skills:
    post:
      tags:
      - "skills"
      summary: "Create new skill"
      operationId: "newSkill"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body required skill object"
        required: true
        schema:
          $ref: '#/definitions/CreateSkill'
      responses:
        200:
          description: "Successed created skill"
        422:
          description: "Invalid parameter exception"
        500:
          description: "Server unexpected error"
    put:
      tags:
      - "skills"
      summary: "Update skill by id in body"
      operationId: "updateSkill"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body required skill object"
        required: true
        schema:
          $ref: '#/definitions/UpdateSkill'
      responses:
        200:
          description: "Successed updated skill"
        404:
          description: "Skill not found"
        422:
          description: "Invalid parameter exception"
        500:
          description: "Server unexpected error"
    get:
      tags:
      - "skills"
      summary: "Get all skills"
      operationId: "getSkills"
      produces:
      - "application/json" 
      responses:
        200:
          description: "Succeed operation"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/SkillResponse'
        204:
          description: "Succeed operation but empty response"
  /skills/{id}:
    get:
      tags: 
      - skills
      summary: "Get skill by id" 
      operationId: "getSkillById"
      produces: 
      - "application/json"
      parameters: 
      - name: "id"
        in: "path"
        description: "ID of skill mapToDbAbility return"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/SkillResponse'
        404:
          description: "Skill not found"
    delete:
      tags: 
      - skills
      summary: "Delete skill by id" 
      operationId: "deleteSkillById"
      parameters: 
      - name: "id"
        in: "path"
        description: "ID of skill mapToDbAbility return"
        required: true
        type: "string"
      responses:
        200:
          description: "Skill deleted"
        404:
          description: "Skill not found"
        500:
          description: "Unexpected server error"
  /abilities:
    post:
      tags:
      - "abilities"
      summary: "Create new ability"
      operationId: "newAbility"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body required ability object"
        required: true
        schema:
          $ref: '#/definitions/CreateAbility'
      responses:
        200:
          description: "Successed created ability"
        422:
          description: "Invalid parameter exception"
        500:
          description: "Server unexpected error"
    put:
      tags:
      - "abilities"
      summary: "Update ability by id in body"
      operationId: "updateAbility"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body required ability object"
        required: true
        schema:
          $ref: '#/definitions/UpdateAbility'
      responses:
        200:
          description: "Successed updated ability"
        404:
          description: "Ability not found"
        422:
          description: "Invalid parameter exception"
        500:
          description: "Server unexpected error"
    get:
      tags:
      - "abilities"
      summary: "Get all abilities"
      operationId: "getAbilities"
      produces:
      - "application/json"
      responses:
        200:
          description: "Succeed operation"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/AbilityResponse'
        204:
          description: "Succeed operation but empty response"
  /abilities/{id}:
    get:
      tags:
      - "abilities"
      summary: "Get ability by id"
      operationId: "getAbilityById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of ability mapToDbAbility return"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/AbilityResponse'
        404:
          description: "Ability not found"
    delete:
      tags:
      - "abilities"
      summary: "Delete ability by id"
      operationId: "deleteAbilityById"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of ability mapToDbAbility return"
        required: true
        type: "string"
      responses:
        200:
          description: "Ability deleted"
        404:
          description: "Ability not found"
        500:
          description: "Unexpected server error"
  /items:
    post:
      tags:
      - "items"
      summary: "Create new item"
      operationId: "newItem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body required item object"
        required: true
        schema:
          $ref: '#/definitions/CreateItem'
      responses:
        200:
          description: "Successed created item"
        422:
          description: "Invalid parameter exception"
        500:
          description: "Server unexpected error"
    put:
      tags:
      - "items"
      summary: "Update item by id in body"
      operationId: "updateItem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body required item object"
        required: true
        schema:
          $ref: '#/definitions/UpdateItem'
      responses:
        200:
          description: "Successed updated item"
        404:
          description: "Item not found"
        422:
          description: "Invalid parameter exception"
        500:
          description: "Server unexpected error"
    get:
      tags:
      - "items"
      summary: "Get all items"
      operationId: "getItems"
      produces:
      - "application/json"
      responses:
        200:
          description: "Succeed operation"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ItemResponse'
        204:
          description: "Succeed operation but empty response"
  /items/{id}:
    get:
      tags:
      - "items"
      summary: "Get item by id"
      operationId: "getItemById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of item mapToDbAbility return"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/ItemResponse'
        404:
          description: "Item not found"
    delete:
      tags:
      - "items"
      summary: "Delete item by id"
      operationId: "deleteItemById"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of item mapToDbAbility return"
        required: true
        type: "string"
      responses:
        200:
          description: "Item deleted"
        404:
          description: "Item not found"
        500:
          description: "Unexpected server error"
...
definitions:
  CreateSkill:
    type: "object"
    required:
    - name
    - trait
    - type
    properties:
      name:
        type: "string"
      description:
        type: "string"
      trait:
        $ref: '#/definitions/TraitsEnum'
      type:
        type: "string"
        enum:
        - BASIC
        - ADVANCE
  UpdateSkill:
    type: "object"
    required:
    - id
    - description
    - name
    - trait
    - type
    properties:
      id:
        type: "string"
      name:
        type: "string"
      description:
        type: "string"
      trait:
        $ref: '#/definitions/TraitsEnum'
      type:
        type: "string"
        enum:
        - BASIC
        - ADVANCE
  SkillResponse:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      description:
        type: "string"
      trait:
        $ref: '#/definitions/TraitsEnum'
      type:
        type: "string"
        enum:
        - BASIC
        - ADVANCE
  CreateAbility:
    type: "object"
    required:
    - name
    - description
    properties:
      name:
        type: "string"
      description:
        type: "string"
  UpdateAbility:
    type: "object"
    required:
    - id
    - description
    - name
    properties:
      id:
        type: "string"
      name:
        type: "string"
      description:
        type: "string"
  AbilityResponse:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      description:
        type: "string"
  CreateItem:
    type: "object"
    required:
    - name
    properties:
      name:
        type: "string"
      price:
        type: "number"
      description:
        type: "string"
      availability:
        $ref: '#/definitions/Availability'
  UpdateItem:
    type: "object"
    required:
    - id
    properties:
      id:
        type: "string"
      name:
        type: "string"
      price:
        type: "number"
      description:
        type: "string"
      availability:
        $ref: '#/definitions/Availability'
  ItemResponse:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      price:
        type: "number"
      description:
        type: "string"
      availability:
        $ref: '#/definitions/Availability'
...
  TraitsEnum:
    type: string
    description: "Value 'other' should be only used in mapToDbAbility operate with skills."
    enum: 
      - WW
      - US
      - K
      - ODP
      - ZR
      - INT
      - SW
      - ODG
      - A
      - ZYW
      - S
      - WT
      - SZ
      - MAG
      - PO
      - PP
      - OTHER
...
  Availability:
    type: "string"
    enum: [rare, little, average, high]
...